Refactor books functionality to `BookStorage` class
