+++ main.rb
@@ -1,10 +1,4 @@
 #!/usr/bin/env ruby
-require './person'
-require './student'
-require './rentals'
-require './book'
-require './teacher'
-require './classroom'
 require './app.rb'
 
 def main
